small rename of errors
